+++ source/Nevermore.IntegrationTests/Model/BrandA.cs
@@ -1,0 +1,8 @@
+namespace Nevermore.IntegrationTests.Model
+{
+    public class BrandA : Brand
+    {
+        public const string BrandType = "BrandA";
+        public override string Type => BrandType;
+    }
+}
