+++ source/Nevermore.IntegrationTests/Model/BrandB.cs
@@ -1,0 +1,8 @@
+namespace Nevermore.IntegrationTests.Model
+{
+    public class BrandB : Brand
+    {
+        public const string BrandType = "BrandB";
+        public override string Type => BrandType;
+    }
+}
