+++ source/Nevermore.IntegrationTests/Model/BrandToTestSerialization.cs
@@ -2,12 +2,11 @@ using Nevermore.Contracts;
 
 namespace Nevermore.IntegrationTests.Model
 {
-    public class Product : IDocument
+    public class BrandToTestSerialization : IDocument
     {
         public string Id { get; set; }
         public string Name { get; set; }
-        public decimal Price { get; set; }
 
-        public ProductType Type { get; set; } = ProductType.Normal;
+        public string JSON { get; set; }
     }
 }
