+++ source/Nevermore.IntegrationTests/Model/BrandToTestSerializationMap.cs
@@ -1,0 +1,14 @@
+using Nevermore.Mapping;
+
+namespace Nevermore.IntegrationTests.Model
+{
+    public class BrandToTestSerializationMap : DocumentMap<BrandToTestSerialization>
+    {
+        public BrandToTestSerializationMap()
+        {
+            TableName = "Brand";
+            Column(x => x.Name);
+            Column(x => x.JSON);
+        }
+    }
+}
