+++ source/Nevermore.IntegrationTests/Model/DodgyProduct.cs
@@ -1,0 +1,12 @@
+namespace Nevermore.IntegrationTests.Model
+{
+    public class DodgyProduct : Product
+    {
+        public DodgyProduct()
+        {
+            Type = ProductType.Dodgy;
+        }
+
+        public decimal Tax { get; set; }
+    }
+}
