+++ source/Nevermore.IntegrationTests/Model/Endpoint.cs
@@ -1,0 +1,9 @@
+namespace Nevermore.IntegrationTests.Model
+{
+    public abstract class Endpoint
+    {
+        public string Name { get; set; }
+
+        public abstract string Type { get; }
+    }
+}
