+++ source/Nevermore.IntegrationTests/Model/PassiveTentacleEndpoint.cs
@@ -1,0 +1,7 @@
+namespace Nevermore.IntegrationTests.Model
+{
+    public class PassiveTentacleEndpoint : Endpoint
+    {
+        public override string Type => "PassiveTentacle";
+    }
+}
