+++ source/Nevermore.IntegrationTests/Model/ProductToTestSerializationMap.cs
@@ -1,0 +1,15 @@
+using Nevermore.Mapping;
+
+namespace Nevermore.IntegrationTests.Model
+{
+    public class ProductToTestSerializationMap : DocumentMap<ProductToTestSerialization>
+    {
+        public ProductToTestSerializationMap()
+        {
+            TableName = "Product";
+            Column(x => x.Name);
+            Column(x => x.Type);
+            Column(x => x.JSON);
+        }
+    }
+}
