+++ source/Nevermore.IntegrationTests/Model/ProductType.cs
@@ -1,0 +1,9 @@
+namespace Nevermore.IntegrationTests.Model
+{
+    public enum ProductType
+    {
+        Normal,
+        Special,
+        Dodgy
+    }
+}
