+++ source/Nevermore.IntegrationTests/Model/SpecialProduct.cs
@@ -1,0 +1,12 @@
+namespace Nevermore.IntegrationTests.Model
+{
+    public class SpecialProduct : Product
+    {
+        public SpecialProduct()
+        {
+            Type = ProductType.Special;
+        }
+
+        public string BonusMaterial { get; set; }
+    }
+}
