+++ source/Nevermore.IntegrationTests/Model/SpecialProductMap.cs
@@ -1,0 +1,11 @@
+namespace Nevermore.IntegrationTests.Model
+{
+    public class SpecialProductMap : ProductMap<SpecialProduct>
+    {
+        public SpecialProductMap()
+        {
+            TableName = typeof(Product).Name;
+            Column(m => m.BonusMaterial).IsNullable = true;
+        }
+    }
+}
